implemented get and update methods

+++ models.go
@@ -3,6 +3,6 @@ package main
 type Product struct {
 	ProductId          int64  `json:"product_id"`
 	ProductName        string `json:"product_name"`
-	ProductQuantity    int32  `json:"product_quantity"`
+	ProductQuantity    int64  `json:"product_quantity"`
 	ProductDescription string `json:"product_description"`
 }
